pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoshelf"
version = "0.1.0"
description = "Offline-first folder organizer with optional Anthropic planning and PySide6 GUI"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Cheol-H-Jeong" }]
dependencies = [
  "loguru>=0.7,<1",
  "pydantic>=2.7,<3",
  "sqlalchemy>=2.0,<3",
]

[project.optional-dependencies]
gui = ["PySide6>=6.8,<7"]
parsers = [
  "pypdf>=5,<6",
  "python-pptx>=1,<2",
  "openpyxl>=3.1,<4",
  "xlrd>=2,<3",
  "python-docx>=1,<2",
  "olefile>=0.47,<1",
]
llm = ["anthropic>=0.51,<1"]
dev = [
  "pytest>=8,<9",
  "ruff>=0.11,<1",
  "pyright>=1.1.390,<2",
]

[project.scripts]
autoshelf = "autoshelf.__main__:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["autoshelf*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
